fix version nodes being replaced by interval; refactor node icon a little
