+++ packages/frinx-device-topology/src/components/node-icons/node-icon.helpers.ts
